refactor: modularize migrations by feature
