+++ src/features/adventure/migrations.js
@@ -1,0 +1,1 @@
+export const migrations = [];
