+++ src/features/proficiency/migrations.js
@@ -1,0 +1,28 @@
+import { WEAPONS } from '../weaponGeneration/data/weapons.js';
+
+export const migrations = [
+  save => {
+    if(!save.proficiency){
+      if(save.proficiencies){
+        save.proficiency = {};
+        for(const [k,v] of Object.entries(save.proficiencies)){
+          save.proficiency[k] = typeof v === 'object' ? (v.level || 0) : v;
+        }
+        delete save.proficiencies;
+      }else{
+        save.proficiency = {};
+      }
+    }
+  },
+  save => {
+    if (save.proficiency && typeof save.proficiency === 'object') {
+      const converted = {};
+      for (const [key, val] of Object.entries(save.proficiency)) {
+        const weapon = WEAPONS[key];
+        const typeKey = weapon?.proficiencyKey || key;
+        converted[typeKey] = (converted[typeKey] || 0) + val;
+      }
+      save.proficiency = converted;
+    }
+  }
+];
