+++ src/features/sect/migrations.js
@@ -1,0 +1,7 @@
+export const migrations = [
+  save => {
+    if(!save.buildings){
+      save.buildings = {};
+    }
+  }
+];
